Changes visibility of mapfield from Uint16 to Uint64
There's a really big performance hit for accessing the bitfield for
visibility of a mapfield. I was hoping to improve that by making the
visibility states byte-aligned, but it seems that does not really help
in a significant way.

It probably should be 16- or 32bit aligned, but then the whole thing
wouldn't fit into one register. That alone is not so bad, but would mean
changing a lot more of the visibility code, which I'd rather not do at
this point.

Also, this commit fixed a bug I previously introduced by changing temp3 to
'char', but at one point that variable was used to store the visibility of
a field. Now it uses wchar_t, which should be right for all supported
compilers.

+++ source/actions/changeview.h
@@ -29,7 +29,7 @@
 
 class ChangeView : public GameAction {
    public:
-      typedef map<MapCoordinate,int> ViewState;
+      typedef map<MapCoordinate,Uint64> ViewState;
       ChangeView( GameMap* gamemap, const ViewState& state );
